pyproject.toml
[tool.poetry]
name = "irtracking"
version = "0.1.0"
description = "Python app for Low Cost Motion Capture"
authors = ["IEatCodeDaily <raisalpwardana@gmail.com>"]
license = "MIT License"
readme = "README.md"
packages = [{include = "irtracking"}]

[tool.poetry.dependencies]
python = "~3.10.11"
numpy = "^2.2.0"
matplotlib = "^3.9.3"
pyvista = "0.44.2"
panel = "^1.5.4"
pycolmap = "^3.11.1"
h5py = "^3.12.1"
ffmpeg-python = "^0.2.0"
scipy = "^1.14.1"
rerun-sdk = "^0.21.0"


[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"
jupyter = "^1.1.1"
ipywidgets = "7.7.1"
ipyvtklink = "0.2.3"
panel = "^1.5.4"
trame = "^3.7.1"
trame-client = "^3.5.0"
trame-server = "^3.2.3"
trame-vtk = "^2.8.12"
trame-vuetify = "^2.7.2"
pytest = "^8.3.4"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
pytest = "^8.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


